+++ core/src/main/java/org/eqasim/core/scenario/cutter/network/NetworkCutter.java
@@ -8,7 +8,8 @@ import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.eqasim.core.scenario.cutter.extent.ScenarioExtent;
 import org.eqasim.core.scenario.cutter.extent.ShapeScenarioExtent;
 import org.matsim.api.core.v01.Id;
@@ -36,7 +37,7 @@ import org.matsim.pt.transitSchedule.api.TransitRoute;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
 
 public class NetworkCutter {
-	private final static Logger log = Logger.getLogger(NetworkCutter.class);
+	private final static Logger log = LogManager.getLogger(NetworkCutter.class);
 
 	private final ScenarioExtent extent;
 	private final MinimumNetworkFinder minimumNetworkFinder;